Reorganize modules to clarify and compartmentalize roles.
